+++ CHANGELOG.md
@@ -2,13 +2,15 @@
 
 ## [Unreleased]
 
-### Changed
+### Added
 
-- Improve REPL demo to follow Block Kit Builder updates ([#77](https://github.com/speee/jsx-slack/pull/77))
+- Add (an experimental) `<Home>` container component for home tab ([#75](https://github.com/speee/jsx-slack/issues/75), [#78](https://github.com/speee/jsx-slack/pull/78))
+- "Copy to clipboard" button on REPL demo ([#77](https://github.com/speee/jsx-slack/pull/77))
 
 ### Fixed
 
 - Throw an error when using `<File>` in `<Modal>` ([#76](https://github.com/speee/jsx-slack/pull/76))
+- REPL demo can transfer the complete modal JSON to Block Kit Builder ([#77](https://github.com/speee/jsx-slack/pull/77))
 
 ### Deprecated